dumper: Load minidump directly instead of use patternsleuth image

+++ dumper/src/containers.rs
@@ -121,7 +121,7 @@ impl<C: Ctx> Ptr<FName, C> {
         if mem.ue_version() < (4, 22) {
             // wtf :skull_emoji:
             let chunks = self
-                .map(|_| mem.fnamepool().0 as u64)
+                .map(|_| mem.fnamepool().0)
                 .cast::<Ptr<Ptr<Ptr<(), C>, C>, C>>()
                 .read()?;
 
@@ -167,9 +167,7 @@ impl<C: Ctx> Ptr<FName, C> {
             });
         }
 
-        let blocks = self
-            .map(|_| mem.fnamepool().0 as u64 + 0x10)
-            .cast::<Ptr<u8, C>>();
+        let blocks = self.map(|_| mem.fnamepool().0 + 0x10).cast::<Ptr<u8, C>>();
 
         let block_index = (value >> 16) as usize;
         let offset = if case_preserving {
@@ -209,4 +207,4 @@ impl<C: Ctx> Ptr<FName, C> {
 }
 
 #[derive(Debug, Clone, Copy)]
-pub struct PtrFNamePool(pub usize);
+pub struct PtrFNamePool(pub u64);
